Add a proper test suite for transformations.
It sort of violates the test protocol in place from this app's ::Critic days, but I think that's okay for now. It also shows up a problem I was afraid would happen, but this can be solved by adding proper ordering. Or at least loose ordering. Or maybe it's just a bug in the way I'm running transformers.