Make these run the proper transformer.
Later on the .run file will move back inside the .t file to make it simpler to see the correlation between the tested and tester. I may go back to having things run out of a central Transformer/ directory if I get tired of the current arrangement, but we'll see.